+++ src/game_2d/GamePanel.java
@@ -2,13 +2,13 @@ package game_2d;
 
 import entity.Player;
 import event.CollisionChecker;
-import event.GameOverState;
-import event.GameState;
-import event.GameStateManager;
-import event.KeyHandler;
-import event.PauseState;
-import event.PlayState;
-import event.TitleMainState;
+import gamestate.GameOverState;
+import gamestate.GameState;
+import gamestate.GameStateManager;
+import gamestate.KeyHandler;
+import gamestate.PauseState;
+import gamestate.PlayState;
+import gamestate.TitleMainState;
 import event.Wave;
 import sound.SoundManager;
 import tile.TileManager;
