+++ src/gamestate/PlayState.java
@@ -2,28 +2,28 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package event;
+package gamestate;
 
 /**
  *
  * @author HP
  */
-public class TitleMainState implements GameState {
+public class PlayState implements GameState {
     private GameStateManager gsm;
-    
-    public TitleMainState(GameStateManager gsm){
+
+    public PlayState(GameStateManager gsm) {
         this.gsm = gsm;
     }
-    
+
     @Override
-    public void changeState(int commandNum){
-        switch (commandNum){
-            case 0 -> gsm.setState(new TitleTutorialState(gsm));
-            case 1 -> System.exit(0);
-        }
+    public void changeState(int commandNum) {
+        gsm.setState(new PauseState(gsm));
+        gsm.getGamePanel().getPlayer().resetFirstFrame();
     }
+
     @Override
-    public GameState getGameState(){
+    public GameState getGameState() {
         return this;
     }
 }
+
