+++ src/gamestate/TitleTutorialState.java
@@ -2,24 +2,25 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package event;
-
-import sound.SoundManager;
+package gamestate;
 
 /**
  *
  * @author HP
  */
-public class PauseState implements GameState {
+public class TitleTutorialState implements GameState {
     private GameStateManager gsm;
 
-    public PauseState(GameStateManager gsm) {
+    public TitleTutorialState(GameStateManager gsm) {
         this.gsm = gsm;
     }
 
     @Override
     public void changeState(int commandNum) {
-        gsm.getGamePanel().restartGame();
+        switch (commandNum) {
+            case 0 -> gsm.setState(new TitleDifficultyState(gsm));
+            case 1 -> gsm.setState(new TitleMainState(gsm));
+        }
     }
 
     @Override
